refactor: 재사용되지 않는 ShowLike 컴포넌트 통합
